% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmarg.R
\name{bcmarg}
\alias{bcmarg}
\title{Marginal Model of the Mixed Effect Model with the Box-Cox Transformation.}
\usage{
bcmarg(formula, data, time = NULL, id = NULL, structure = "UN")
}
\arguments{
\item{formula}{a two-sided linear formula object describing the model, with
the response on the left of a ~ operator and the terms, separated by +
operators, on the right.}

\item{data}{a data frame in which to interpret the variables named in the
formula.}

\item{time}{time variable name for repeated measurements. Default is NULL.}

\item{id}{subject id variable name for repeated measurements. Default is
NULL.}

\item{structure}{specify the covariance structure from c("UN", "CS",
"AR(1)"). Default is "UN".}
}
\value{
bcmarg returns a list including following conponents for the
marginal model parameter inference:
\describe{
  \item{\code{lambda}}{estimate of a transformation parameter.}
  \item{\code{beta}}{estimate of a regression parameter vector.}
  \item{\code{alpha}}{estimate of a scale parameter vector.}
  \item{\code{V}}{variace-covariance matrix for any subject with no missing
        values.}
  \item{\code{betainf}}{inference results for beta under the assumption
        that lambda is known.}
  \item{\code{Vtheta.mod}}{model-based variance-covariance matrix for MLE of
        a vector of all parameters.}
  \item{\code{Vtheta.rob}}{robust variance-covariance matrix for MLE of
        a vector of all parameters.}
  \item{\code{lik}}{maximized liklihood.}
  \item{\code{adj.prm}}{small-sample adjustment parameter used in
        \code{\link{bcmmrm}}: c(number of subjects, number of
        completed subjects, number of outcome observations, number of
        missing observations).}
  \item{\code{glsresult}}{results for \code{gls} (or \code{lm} when time
        and id are not specified) function on the transformed scale. }
}
}
\description{
\code{bcmarg} returns the inference resutls the parameters of
  the marginal model of the linear mixed effect model with the Box-Cox
  transformation proposed by Maruo et al. (2017). If time and id are not
  specified, inference results reduce to the results for the context of
  linear regression model proveded by Maruo et al. (2015).
}
\examples{
 data(aidscd4)
 bcmarg(cd4 ~ as.factor(treatment), aidscd4, weekc, id, "AR(1)")

}
\references{
\itemize{
  \item Maruo, K., Isogawa, N., Gosho, M. (2015). Inference of median
        difference based on the Box-Cox model in randomized clinical trials.
        \emph{Statistics in Medicine}, 34, 1634-1644.\cr
  \item Maruo, K., Yamaguchi, Y., Noma, H., Gosho, M. (2017). Interpretable
        inference on the mixed effect model with the Box-Cox transformation.
        \emph{Statistics in Medicine}, 36, 2420-2434.
}
}
\seealso{
\code{\link{bcmmrm}}
}
