% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmixed-package.R
\docType{package}
\name{bcmixed-package}
\alias{bcmixed}
\alias{bcmixed-package}
\title{bcmixed: Mixed Effect Model with the Box-Cox Transformation}
\description{
Inference on the marginal model of the mixed effect model with the Box-Cox transformation and on the model median differences between treatment groups for longitudinal randomized clinical trials. These statistical methods are proposed by Maruo et al. (2017) \doi{10.1002/sim.7279}.
}
\author{
\strong{Maintainer}: Kazushi Maruo \email{kazushi.maruo@gmail.com}

Authors:
\itemize{
  \item Ryota Ishii
}

Other contributors:
\itemize{
  \item Yusuke Yamaguchi [contributor]
  \item Masahiko Gosho [contributor]
}

}
\keyword{internal}
