% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmixed.R
\docType{package}
\name{bcmixed}
\alias{bcmixed}
\title{bcmixed: A package for mixed effect models with the Box-Cox transformations.}
\description{
The bcmixed package provides three categories of important functions:
bcreg, bcmarg, and bcmmrm.
}
\details{
\tabular{ll}{
Package:\tab bcmixed\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-04-12\cr
}
}
\author{
Kazushi Maruo
}
\keyword{package}
