% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmixed.R
\docType{package}
\name{bcmixed}
\alias{bcmixed}
\title{bcmixed: A package for mixed effect models with the Box-Cox transformations.}
\description{
The bcmixed package provides two categories of important functions:
\code{\link{bcmarg}} and \code{\link{bcmmrm}}. The \code{\link{bcmarg}}
function provides inferences on the marginal model of the mixed effect
model with the Box-Cox transformation and the \code{\link{bcmmrm}}
provides inferences on the model median differences between treatment
groups for longitudinal randomized clinical trials. These statistical
methods are proposed by Maruo et al. (2017,
\url{https://doi.org/10.1002/sim.7279}).
}
\details{
\tabular{ll}{
Package:\tab bcmixed\cr
Type: \tab Package\cr
Version: \tab 0.1.3\cr
Date: \tab 2020-02-22\cr
}
}
\author{
Kazushi Maruo, Ryota Ishii, Yusuke Yamaguchi, Masahiko Gosho
}
\keyword{package}
