% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmmrm.R
\name{bcmmrm}
\alias{bcmmrm}
\title{Model Median Inference for Longitudinal Data in Randomized Clinical Trials.}
\usage{
bcmmrm(outcome, group, data, time = NULL, id = NULL, timepoint = NULL,
  covv = NULL, cfactor = NULL, structure = "UN")
}
\arguments{
\item{outcome}{positive outcome (dependent) variable.}

\item{group}{treatment group variable.}

\item{data}{a data frame which includes outcome, group, time, id, and
specified covariate variables.}

\item{time}{time variable for repeated measurements.}

\item{id}{subject id variable for repeated measurements.}

\item{timepoint}{analysis time point.}

\item{covv}{covariate variables. default is NULL.}

\item{cfactor}{nominal variable indicator for covaraite variables. nominal
variable: 1, continuous variable: 0. default is NULL.}

\item{structure}{specify the covariance structure from c("UN", "CS",
"AR(1)"). default is "UN"}
}
\value{
bcmmrm returns a list including following components for the model
  median inference. \describe{
    \item{\code{median}}{inrefence results for the model medians for all
          groups.}
    \item{\code{meddif.mod}}{model-based inference results for the model
          median differences between all pairs of groups.}
    \item{\code{meddif.rob}}{robust inference results for the model median
          differences between all pairs of groups.}
    \item{\code{meddif.mod.adj}}{model-based inference results for the model
          median differences between all pairs of groups with the empirical
          adjustment.}
    \item{\code{meddif.rob.adj}}{robust inference results for the model
          median differences between all pairs of groups with the empirical
          adjustment.}
    \item{\code{lambda}}{estimate of a transformation parameter.}
    \item{\code{R}}{correlation matrix for any subject with no missing
          values.}
    \item{\code{betainf}}{inference results for beta under the assumption
          that lambda is known.}
    \item{\code{inf.marg}}{result of \code{\link{bcmarg}} function.}
    \item{\code{outdata}}{data frame where the transformed outcome
          (\code{ytr}) and the fitted value (\code{ytr.fitted}) on the
          transformed scale are added ito input data.}
  }
}
\description{
\code{bcmmrm} provides inference results for the model median differences
between treatment groups proposed by Maruo et al. (2017), which focuses on
continuous and positive longitudinally observed outcomes and a situation
where the efficacy of some treatments is compared based on a randomized,
parallel group clinical trial. If time and id are not specified, inference
results reduce to the results for the context of linear regression model
proveded by Maruo et al. (2015).
}
\note{
If baseline observation for the outcome variable is available, Box-Cox
  transformed baseline should be included as a covariate for accuracy of
  estimation.\cr Although this function can be applied to non-randomized
  trial data, performances of the above approach have not evaluated in
  context of non-randomized trials.
}
\examples{
data(aidscd4)
# covariates: no covariate, covariance structure: CS structure
bcmmrm(cd4, treatment, aidscd4, weekc, id, 32, structure="CS")

# covariate: Box-Cox transformed baseline,age (continuous variables),
# and sex (nominal variable), covariance structure: UN structure
lmd.bl <- bcreg(cd4.bl ~ 1, aidscd4[aidscd4$weekc==8, ])$lambda
aidscd4$cd4.bl.tr <- (aidscd4$cd4.bl^lmd.bl-1)/lmd.bl
bcmmrm(cd4, treatment, aidscd4, weekc, id, 32, c("cd4.bl.tr"), c(0), structure = "AR(1)")

}
\references{
\itemize{
  \item Maruo, K., Isogawa, N., Gosho, M. (2015). Inference of median
        difference based on the Box-Cox model in randomized clinical trials.
        \emph{Statistics in Medicine}, 34, 1634-1644.\cr
  \item Maruo, K., Yamaguchi, Y., Noma, H., Gosho, M. (2017). Interpretable
        inference on the mixed effect model with the Box-Cox transformation.
        \emph{Statistics in Medicine}, early view (DOI: 10.1002/sim.7279).
}
}
\seealso{
\code{\link{bcmarg}}
}
