% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcreg.R
\name{bcreg}
\alias{bcreg}
\title{Linear regression model with the Box-Cox Transformation.}
\usage{
bcreg(formula, data)
}
\arguments{
\item{formula}{a two-sided linear formula object describing the model, with
the response on the left of a ~ operator and the terms, separated by +
operators, on the right.}

\item{data}{a data frame in which to interpret the variables named in the
formula.}
}
\value{
LmeMargBc returns a list including following conponents:
\describe{
  \item{\code{lambda}}{estimate of a transformation parameter.}
  \item{\code{beta}}{estimate of a regression parameter vector.}
  \item{\code{sigma}}{estimate of a scale parameter.}
  \item{\code{betainf}}{inference results for beta under the
        assumption that lambda is known.}
  \item{\code{lik}}{maximized liklihood.}
  \item{\code{lmresult}}{results of \code{lm} function on the
        transformed scale}
}
}
\description{
\code{bcreg} returns the maximum likelihood estimates for parameters of
the linear regression models with the Box-Cox transformation (Box and Cox,
1964).
}
\examples{
  data(aidscd4)
  #Transformation of baseline observation for aid.cd4 data
  bcreg(cd4.bl ~ 1, aidscd4[aidscd4$weekc == 8, ])

}
\references{
Box, G.E.P. and Cox, D.R. (1964). An analysis of transformations
(with discussion). Journals of the Royal Statistical Society, Series B, 26,
211-246.
}
